Add base event for quote vote added
